refactor: add `log.ts` separated from `index.ts`
